refactor: move models back to resource & datasource
